backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose==3.5.0
httpx==0.28.1
email-validator==2.2.0
